# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Intelligent scheduling system for UK personal trainers.",
        "title": "PT Scheduler API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "PT Scheduler Support",
            "email": "support@pt-scheduler.io"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/billing/charge": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually triggers the monthly billing charge for a client. Idempotent — safe to call multiple times for the same billing period.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Trigger monthly charge",
                "parameters": [
                    {
                        "description": "Charge details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_billing.chargeBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_billing.Payment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already charged for this billing period",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ValidationErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/payments/mandate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Starts a GoCardless redirect flow for Direct Debit setup. FRONTEND: redirect the client to the returned redirect_url.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Start GoCardless mandate flow",
                "parameters": [
                    {
                        "description": "Client ID and redirect URI",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_billing.MandateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_billing.MandateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ValidationErrorResponse"
                        }
                    }
                }
            }
        },
        "/payments/mandate/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Finalises a GoCardless redirect flow after the client has authorised the Direct Debit mandate.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Complete GoCardless mandate flow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GoCardless redirect flow ID",
                        "name": "redirect_flow_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Client ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_billing.completeMandateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_billing.MandateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payments/setup-intent": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a Stripe SetupIntent so the client can save their card without an immediate charge. FRONTEND: use the returned client_secret to initialise Stripe Elements.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Create Stripe SetupIntent",
                "parameters": [
                    {
                        "description": "Client details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_billing.setupIntentBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_billing.SetupIntentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ValidationErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/webhooks/gocardless": {
            "post": {
                "description": "Receives and verifies GoCardless webhook events. Authenticated by Webhook-Signature header HMAC, not JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "GoCardless webhook receiver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GoCardless webhook signature",
                        "name": "Webhook-Signature",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/stripe": {
            "post": {
                "description": "Receives and verifies Stripe webhook events. Authenticated by Stripe-Signature header HMAC, not JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Stripe webhook receiver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stripe webhook signature",
                        "name": "Stripe-Signature",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "internal_billing.MandateRequest": {
            "type": "object",
            "required": [
                "client_id",
                "redirect_uri"
            ],
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "redirect_uri": {
                    "type": "string"
                }
            }
        },
        "internal_billing.MandateResponse": {
            "type": "object",
            "properties": {
                "flow_id": {
                    "type": "string"
                },
                "redirect_url": {
                    "description": "FRONTEND: redirect client to this URL",
                    "type": "string"
                }
            }
        },
        "internal_billing.Payment": {
            "type": "object",
            "properties": {
                "amount_pence": {
                    "type": "integer"
                },
                "billing_month": {
                    "type": "integer"
                },
                "billing_year": {
                    "type": "integer"
                },
                "client_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "provider_ref": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_billing.SetupIntentResponse": {
            "type": "object",
            "properties": {
                "client_secret": {
                    "type": "string"
                },
                "setup_intent_id": {
                    "type": "string"
                }
            }
        },
        "internal_billing.chargeBody": {
            "type": "object",
            "required": [
                "amount_pence",
                "billing_month",
                "billing_year",
                "client_id",
                "provider"
            ],
            "properties": {
                "amount_pence": {
                    "type": "integer",
                    "minimum": 100
                },
                "billing_month": {
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1
                },
                "billing_year": {
                    "type": "integer",
                    "minimum": 2024
                },
                "client_id": {
                    "type": "string"
                },
                "provider": {
                    "type": "string",
                    "enum": [
                        "stripe",
                        "gocardless"
                    ]
                }
            }
        },
        "internal_billing.completeMandateBody": {
            "type": "object",
            "required": [
                "client_id"
            ],
            "properties": {
                "client_id": {
                    "type": "string"
                }
            }
        },
        "internal_billing.setupIntentBody": {
            "type": "object",
            "required": [
                "client_id",
                "email",
                "full_name"
            ],
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer \" followed by your access token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}